+++ crates/egui_demo_lib/src/demo/mod.rs
@@ -17,6 +17,7 @@ pub mod frame_demo;
 pub mod highlighting;
 pub mod interactive_container;
 pub mod misc_demo_window;
+pub mod modals;
 pub mod multi_touch;
 pub mod paint_bezier;
 pub mod painting;