+++ crates/egui_kittest/src/lib.rs
@@ -119,7 +119,7 @@ impl<'a, State> Harness<'a, State> {
     ///     });
     /// }, checked);
     ///
-    /// harness.get_by_name("Check me!").click();
+    /// harness.get_by_label("Check me!").click();
     /// harness.run();
     ///
     /// assert_eq!(*harness.state(), true);
@@ -144,7 +144,7 @@ impl<'a, State> Harness<'a, State> {
     ///     ui.checkbox(checked, "Check me!");
     /// }, checked);
     ///
-    /// harness.get_by_name("Check me!").click();
+    /// harness.get_by_label("Check me!").click();
     /// harness.run();
     ///
     /// assert_eq!(*harness.state(), true);
@@ -248,6 +248,25 @@ impl<'a, State> Harness<'a, State> {
     /// Access the state mutably.
     pub fn state_mut(&mut self) -> &mut State {
         &mut self.state
+    }
+
+    /// Press a key.
+    /// This will create a key down event and a key up event.
+    pub fn press_key(&mut self, key: egui::Key) {
+        self.input.events.push(egui::Event::Key {
+            key,
+            pressed: true,
+            modifiers: Default::default(),
+            repeat: false,
+            physical_key: None,
+        });
+        self.input.events.push(egui::Event::Key {
+            key,
+            pressed: false,
+            modifiers: Default::default(),
+            repeat: false,
+            physical_key: None,
+        });
     }
 }